+++ crates/rspack_core/src/compiler/compilation.rs
@@ -39,8 +39,8 @@ use crate::{
     OptimizeDependencyResult,
   },
   BuildContext, BundleEntries, Chunk, ChunkByUkey, ChunkGraph, ChunkGroup, ChunkGroupUkey,
-  ChunkUkey, CodeGenerationResult, CompilerOptions, Dependency, EntryItem, Entrypoint,
-  LoaderRunnerRunner, ModuleDependency, ModuleGraph, ModuleIdentifier, ModuleRule, Msg,
+  ChunkUkey, CodeGenerationResult, CodeGenerationResults, CompilerOptions, Dependency, EntryItem,
+  Entrypoint, LoaderRunnerRunner, ModuleDependency, ModuleGraph, ModuleIdentifier, ModuleRule, Msg,
   NormalModule, NormalModuleFactory, NormalModuleFactoryContext, ProcessAssetsArgs,
   RenderManifestArgs, RenderRuntimeArgs, ResolveKind, Runtime, SharedPluginDriver, Stats,
   VisitedModuleIdentity,
@@ -872,176 +872,6 @@ impl CompilationAsset {
   }
 }
 
-type RuntimeSpec = HashSet<String>;
-type RuntimeKey = String;
-
-#[derive(Default, Debug)]
-enum RuntimeMode {
-  #[default]
-  Empty = 0,
-  SingleEntry = 1,
-  Map = 2,
-}
-
-pub fn is_runtime_equal(a: &RuntimeSpec, b: &RuntimeSpec) -> bool {
-  if a.len() != b.len() {
-    return false;
-  }
-
-  let mut a: Vec<String> = Vec::from_iter(a.iter().cloned());
-  let mut b: Vec<String> = Vec::from_iter(b.iter().cloned());
-
-  a.sort();
-  b.sort();
-
-  a.into_iter().zip(b.into_iter()).all(|(a, b)| a == b)
-}
-
-pub fn get_runtime_key(runtime: RuntimeSpec) -> String {
-  let mut runtime: Vec<String> = Vec::from_iter(runtime.into_iter());
-  runtime.sort();
-  runtime.join("\n")
-}
-
-#[derive(Default, Debug)]
-pub struct RuntimeSpecMap {
-  mode: RuntimeMode,
-  map: HashMap<RuntimeKey, CodeGenerationResult>,
-
-  single_runtime: Option<RuntimeSpec>,
-  single_value: Option<CodeGenerationResult>,
-}
-
-impl RuntimeSpecMap {
-  fn size(&self) -> usize {
-    let mode = self.mode as usize;
-
-    if mode <= 1 {
-      mode
-    } else {
-      self.map.len()
-    }
-  }
-
-  fn get(&self, runtime: &RuntimeSpec) -> Option<&CodeGenerationResult> {
-    match self.mode {
-      RuntimeMode::Empty => None,
-      RuntimeMode::SingleEntry => {
-        if let Some(single_runtime) = self.single_runtime.as_ref() && is_runtime_equal(single_runtime, runtime) {
-          self.single_value.as_ref()
-        } else {
-          None
-        }
-      }
-      RuntimeMode::Map => self.map.get(&get_runtime_key(runtime.clone())),
-    }
-  }
-
-  fn set(&mut self, runtime: RuntimeSpec, value: CodeGenerationResult) {
-    match self.mode {
-      RuntimeMode::Empty => {
-        self.mode = RuntimeMode::SingleEntry;
-        self.single_runtime = Some(runtime);
-        self.single_value = Some(value);
-      }
-      RuntimeMode::SingleEntry => {
-        if let Some(single_runtime) = self.single_runtime.as_ref() && is_runtime_equal(single_runtime, &runtime) {
-          self.single_value = Some(value);
-        } else {
-          self.mode = RuntimeMode::Map;
-
-          let single_runtime = self.single_runtime.take().expect("Expected single runtime exists");
-          let single_value = self.single_value.take().expect("Expected single value exists");
-
-          self.map.insert(get_runtime_key(single_runtime), single_value);
-        }
-      }
-      RuntimeMode::Map => {
-        self.map.insert(get_runtime_key(runtime), value);
-      }
-    }
-  }
-
-  fn get_values(&self) -> Vec<&CodeGenerationResult> {
-    match self.mode {
-      RuntimeMode::Empty => vec![],
-      RuntimeMode::SingleEntry => vec![self
-        .single_value
-        .as_ref()
-        .expect("Expected single value exists")],
-      RuntimeMode::Map => self.map.values().collect(),
-    }
-  }
-}
-
-#[derive(Default, Debug)]
-pub struct CodeGenerationResults {
-  map: HashMap<ModuleIdentifier, RuntimeSpecMap>,
-}
-
-impl CodeGenerationResults {
-  pub fn get(
-    &self,
-    module_identifier: &ModuleIdentifier,
-    runtime: Option<&RuntimeSpec>,
-  ) -> Result<&CodeGenerationResult> {
-    if let Some(entry) = self.map.get(module_identifier) {
-      if let Some(runtime) = runtime {
-        entry.get(runtime).ok_or_else(|| {
-          Error::InternalError(format!(
-            "Failed to code generation result for {} with runtime {:?}",
-            module_identifier, runtime
-          ))
-        })
-      } else {
-        if entry.size() > 1 {
-          let results = entry.get_values();
-          if results.len() != 1 {
-            return Err(Error::InternalError(format!(
-              "No unique code generation entry for unspecified runtime for {} ",
-              module_identifier,
-            )));
-          }
-
-          return results
-            .first()
-            .copied()
-            .ok_or_else(|| Error::InternalError("Expected value exists".to_string()));
-        }
-
-        entry
-          .get_values()
-          .first()
-          .copied()
-          .ok_or_else(|| Error::InternalError("Expected value exists".to_string()))
-      }
-    } else {
-      Err(Error::InternalError(format!(
-        "No code generation entry for {} (existing entries: {:?})",
-        module_identifier,
-        self.map.keys().collect::<Vec<_>>()
-      )))
-    }
-  }
-
-  fn add(
-    &mut self,
-    module_identifier: ModuleIdentifier,
-    runtime: RuntimeSpec,
-    result: CodeGenerationResult,
-  ) {
-    match self.map.entry(module_identifier) {
-      hashbrown::hash_map::Entry::Occupied(mut record) => {
-        record.get_mut().set(runtime, result);
-      }
-      hashbrown::hash_map::Entry::Vacant(record) => {
-        let mut spec_map = RuntimeSpecMap::default();
-        spec_map.set(runtime, result);
-        record.insert(spec_map);
-      }
-    };
-  }
-}
 #[derive(Debug, Default, Clone)]
 pub struct AssetInfo {
   /// if the asset can be long term cached forever (contains a hash)
